matrix/requirements.txt
matrix-nio==0.24.0
httpx==0.28.1
pydantic-settings==2.6.1
